+++ apps/extension/src/store/middlewares/settingsMiddlewares.ts
@@ -1,0 +1,16 @@
+import { createListenerMiddleware } from "@reduxjs/toolkit";
+import {
+  changeCheckIPInBackground,
+  changeShowPublicIPNotification,
+} from "@reducers/settingsReducer";
+
+export const settingsListener = createListenerMiddleware();
+
+settingsListener.startListening({
+  actionCreator: changeShowPublicIPNotification,
+  effect: async (action, listenerApi) => {
+    if (action.payload) {
+      listenerApi.dispatch(changeCheckIPInBackground(true));
+    }
+  },
+});